feat: upgrade llama-rs

+++ packages/core/src/types.rs
@@ -1,5 +1,4 @@
 use std::sync::mpsc::Sender;
-use napi::bindgen_prelude::{BigInt};
 
 #[napi(object)]
 #[derive(Clone, Debug)]
@@ -38,21 +37,21 @@ pub struct TokenizeResult {
 #[derive(Clone, Debug)]
 pub struct LLamaConfig {
   pub path: String,
-  pub num_ctx_tokens: Option<i32>,
+  pub num_ctx_tokens: Option<i64>,
 }
 
 #[napi(object)]
 #[derive(Clone, Debug)]
 pub struct LLamaInferenceArguments {
   pub n_threads: Option<i32>,
-  pub n_batch: Option<BigInt>,
-  pub top_k: Option<BigInt>,
+  pub n_batch: Option<i64>,
+  pub top_k: Option<i64>,
   pub top_p: Option<f64>,
   pub repeat_penalty: Option<f64>,
   pub temp: Option<f64>,
-  pub seed: Option<BigInt>,
-  pub num_predict: Option<BigInt>,
-  pub repeat_last_n: Option<BigInt>,
+  pub seed: Option<i64>,
+  pub num_predict: Option<i64>,
+  pub repeat_last_n: Option<i64>,
   pub prompt: String,
   pub float16: Option<bool>,
   pub token_bias: Option<String>,
